importer

+++ services/gateway/gateway-api-spec/src/main/java/com/example/gateway/api/model/OptionJournalGWModel.java
@@ -11,17 +11,17 @@ public class OptionJournalGWModel {
     private String accountId;
     private OffsetDateTime date;
     private String stockSymbol;
-    private float stockPrice;
-    private float strikePrice;
+    private double stockPrice;
+    private double strikePrice;
     private OffsetDateTime expiryDate;
-    private float impliedVolatility;
-    private float historicalImpliedVolatility;
-    private float profitProbability;
+    private double impliedVolatility;
+    private double historicalImpliedVolatility;
+    private double profitProbability;
     private int contracts;
-    private float premium;
+    private double premium;
     private ActionGW actionGW;
     private ActionTypeGW actionTypeGW;
-    private float brokerFees;
+    private double brokerFees;
     private String mark;
 
 
@@ -49,15 +49,15 @@ public class OptionJournalGWModel {
         return stockSymbol;
     }
 
-    public float getStockPrice() {
+    public double getStockPrice() {
         return stockPrice;
     }
 
-    public float getImpliedVolatility() {
+    public double getImpliedVolatility() {
         return impliedVolatility;
     }
 
-    public float getHistoricalImpliedVolatility() {
+    public double getHistoricalImpliedVolatility() {
         return historicalImpliedVolatility;
     }
 
@@ -69,7 +69,7 @@ public class OptionJournalGWModel {
         return actionTypeGW;
     }
 
-    public float getBrokerFees() {
+    public double getBrokerFees() {
         return brokerFees;
     }
 
@@ -77,7 +77,7 @@ public class OptionJournalGWModel {
         return pairTransactionId;
     }
 
-    public float getStrikePrice() {
+    public double getStrikePrice() {
         return strikePrice;
     }
 
@@ -85,7 +85,7 @@ public class OptionJournalGWModel {
         return expiryDate;
     }
 
-    public float getProfitProbability() {
+    public double getProfitProbability() {
         return profitProbability;
     }
 
@@ -93,7 +93,7 @@ public class OptionJournalGWModel {
         return contracts;
     }
 
-    public float getPremium() {
+    public double getPremium() {
         return premium;
     }
 
@@ -139,17 +139,17 @@ public class OptionJournalGWModel {
             return this;
         }
 
-        public Builder withStockPrice(float stockPrice) {
+        public Builder withStockPrice(double stockPrice) {
             optionJournalModel.stockPrice = stockPrice;
             return this;
         }
 
-        public Builder withImpliedVolatility(float impliedVolatility) {
+        public Builder withImpliedVolatility(double impliedVolatility) {
             optionJournalModel.impliedVolatility = impliedVolatility;
             return this;
         }
 
-        public Builder withHistoricalImpliedVolatility(float historicalImpliedVolatility) {
+        public Builder withHistoricalImpliedVolatility(double historicalImpliedVolatility) {
             optionJournalModel.historicalImpliedVolatility = historicalImpliedVolatility;
             return this;
         }
@@ -164,7 +164,7 @@ public class OptionJournalGWModel {
             return this;
         }
 
-        public Builder withBrokerFees(float brokerFees) {
+        public Builder withBrokerFees(double brokerFees) {
             optionJournalModel.brokerFees = brokerFees;
             return this;
         }
@@ -174,7 +174,7 @@ public class OptionJournalGWModel {
             return this;
         }
 
-        public Builder withStrikePrice(float strikePrice) {
+        public Builder withStrikePrice(double strikePrice) {
             optionJournalModel.strikePrice = strikePrice;
             return this;
         }
@@ -184,7 +184,7 @@ public class OptionJournalGWModel {
             return this;
         }
 
-        public Builder withProfitProbability(float profitProbability) {
+        public Builder withProfitProbability(double profitProbability) {
             optionJournalModel.profitProbability = profitProbability;
             return this;
         }
@@ -194,7 +194,7 @@ public class OptionJournalGWModel {
             return this;
         }
 
-        public Builder withPremium(float premium) {
+        public Builder withPremium(double premium) {
             optionJournalModel.premium = premium;
             return this;
         }
